requirements.txt
beautifulsoup4==4.10.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
docutils==0.18.1
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
Kivy==2.1.0
Kivy-Garden==0.1.5
MarkupSafe==2.1.1
numpy==1.22.4
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
pyDeprecate==0.3.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
torch==1.11.0
torchmetrics==0.8.1
torchvision==0.12.0
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
